demo(android): unify camera-controls + gesture-editing into camera-gestures (#2239 Batch 1)
Consolidates the retired `camera-controls` and `gesture-editing` demos
into a single `camera-gestures` entry with a top segmented-button toggle:

- **Camera Modes** (default) — Filament `Manipulator` modes (orbit /
  free-flight / map) with a distance slider, an on-screen flight pad
  (FREE_FLIGHT), and the existing Reset Camera action.
- **Node Gestures** — per-node drag / twist / pinch on the damaged
  helmet, with editable / per-axis locks, scale sensitivity, gesture-mode
  indicator, and a live transform readout overlay.

Each sub-mode keeps its own `SceneView` + camera/gesture pipeline — the
two were intentionally distinct and merging them into one viewport would
collapse the "two distinct primitives" story.

### Deep-link aliases

`camera-controls` and `gesture-editing` redirect to `camera-gestures`
through `DeepLinkRouter.DEMO_ID_ALIASES`. Maestro flows in
`.maestro/android/interaction.yaml` keep working (assert "Navigate back"
only).

### Tests

- `DemoInteractionTest.gestureEditing_editableAndReset` — opens unified
  demo, switches to Node Gestures tab, exercises editable + reset flow.
- `DemoSmokeTest.z01_cameraControls_smokeOpen` — points at unified demo
  (default Camera Modes tab still builds the Filament Manipulator).
- `DemoRenderingScreenshotTest.cameraAndGesturesDemo_default_state` —
  renamed, points at new `camera-gestures` slug.
- Stale `cameracontrols_default.png` and `gestureediting_default.png`
  goldens removed; harness will re-capture on first run.

### Quality gates

- ✅ `./gradlew :samples:android-demo:assembleDebug`
- ✅ `./gradlew :sceneview:detekt :arsceneview:detekt :sceneview-core:detekt`
- ✅ `bash samples/android-demo/scripts/collate-demos.sh` (60 fragments)

### Out of scope

- iOS mirror — Android-only per Batch 1 plan.
- Tab pre-selection from alias — follow-up.

Refs #2239 Batch 1.
