+++ changelog.d/2239-batch-1-camera-and-gestures.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified Camera & Gestures demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `camera-controls` and `gesture-editing` demos consolidated into a single `camera-gestures` entry with a segmented-button toggle between Camera Modes (orbit / free-flight / map manipulator + distance slider) and Node Gestures (per-node drag / twist / pinch with editability locks, scale sensitivity, and live transform readout) modes. Existing `sceneview://demo/camera-controls` and `sceneview://demo/gesture-editing` deep links keep working via `DEMO_ID_ALIASES`.
