+++ llms.txt
@@ -4015,9 +4015,8 @@ by `samples/android-demo/scripts/collate-demos.sh` — never edit between the ma
 
 ### Interaction
 
-- `camera-controls` — Camera Controls. Orbit, pan, and free camera modes.
+- `camera-gestures` — Camera & Gestures. Manipulator modes and per-node edit gestures.
 - `collision` — Collision & Hit Test. Hit testing and collision detection.
-- `gesture-editing` — Gesture Editing. Move, scale, and rotate with gestures.
 - `view-node` — ViewNode. Native UI embedded in 3D space.
 
 ### Advanced
