+++ samples/android-demo/src/debug/java/io/github/sceneview/demo/DemoHostActivity.kt
@@ -20,7 +20,7 @@ import io.github.sceneview.demo.demos.ARStreetscapeDemo
 import io.github.sceneview.demo.demos.ARTerrainAnchorDemo
 import io.github.sceneview.demo.demos.AnimationDemo
 import io.github.sceneview.demo.demos.BillboardDemo
-import io.github.sceneview.demo.demos.CameraControlsDemo
+import io.github.sceneview.demo.demos.CameraAndGesturesDemo
 import io.github.sceneview.demo.demos.CollisionDemo
 import io.github.sceneview.demo.demos.CustomMeshDemo
 import io.github.sceneview.demo.demos.DebugOverlayDemo
@@ -28,7 +28,6 @@ import io.github.sceneview.demo.demos.DynamicSkyDemo
 import io.github.sceneview.demo.demos.EnvironmentDemo
 import io.github.sceneview.demo.demos.FogDemo
 import io.github.sceneview.demo.demos.GeometryDemo
-import io.github.sceneview.demo.demos.GestureEditingDemo
 import io.github.sceneview.demo.demos.ImageDemo
 import io.github.sceneview.demo.demos.LightingDemo
 import io.github.sceneview.demo.demos.LinesPathsDemo
@@ -117,8 +116,8 @@ class DemoHostActivity : ComponentActivity() {
             "billboard" -> BillboardDemo(onBack = back)
             "video" -> VideoDemo(onBack = back)
             // Interaction
-            "camera-controls" -> CameraControlsDemo(onBack = back)
-            "gesture-editing" -> GestureEditingDemo(onBack = back)
+            // #2239 Batch 1 — `camera-controls` and `gesture-editing` consolidated into `camera-gestures`.
+            "camera-gestures", "camera-controls", "gesture-editing" -> CameraAndGesturesDemo(onBack = back)
             "collision" -> CollisionDemo(onBack = back)
             "view-node" -> ViewNodeDemo(onBack = back)
             // Advanced
