+++ samples/android-demo/src/main/res/values/strings_demo_camera_and_gestures.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Strings for the unified `camera-gestures` demo (registered by
+     `CameraAndGesturesFragment.kt`). Consolidates the retired
+     `camera-controls` and `gesture-editing` demos (#2239 Batch 1). Old ids
+     continue to deep-link through `DeepLinkRouter.DEMO_ID_ALIASES`. -->
+<resources>
+    <string name="demo_camera_and_gestures_title">Camera &amp; Gestures</string>
+    <string name="demo_camera_and_gestures_subtitle">Manipulator modes and per-node edit gestures</string>
+</resources>
